+++ admin-panel/lib/dependency_injection.dart
@@ -1,13 +1,13 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:conference_admin/core/services/article_service.dart';
-import 'package:conference_admin/features/about/data/repositories/about_repo_impl.dart';
-import 'package:conference_admin/features/about/domain/repositories/about_repo.dart';
-import 'package:conference_admin/features/about/domain/usecases/create_component_uc.dart';
-import 'package:conference_admin/features/about/domain/usecases/delete_component_uc.dart';
-import 'package:conference_admin/features/about/domain/usecases/get_page_by_id_uc.dart';
-import 'package:conference_admin/features/about/domain/usecases/get_pages_uc.dart';
-import 'package:conference_admin/features/about/domain/usecases/update_component_uc.dart';
-import 'package:conference_admin/features/about/presentation/bloc/about_bloc.dart';
+import 'package:conference_admin/features/pages/data/repositories/about_repo_impl.dart';
+import 'package:conference_admin/features/pages/domain/repositories/about_repo.dart';
+import 'package:conference_admin/features/pages/domain/usecases/create_component_uc.dart';
+import 'package:conference_admin/features/pages/domain/usecases/delete_component_uc.dart';
+import 'package:conference_admin/features/pages/domain/usecases/get_page_by_id_uc.dart';
+import 'package:conference_admin/features/pages/domain/usecases/get_pages_uc.dart';
+import 'package:conference_admin/features/pages/domain/usecases/update_component_uc.dart';
+import 'package:conference_admin/features/pages/presentation/bloc/about_bloc.dart';
 import 'package:conference_admin/features/article/data/repositories/article_repo_impl.dart';
 import 'package:conference_admin/features/article/domain/repositories/article_repo.dart';
 import 'package:conference_admin/features/article/domain/usecases/add_article_uc.dart';
@@ -37,11 +37,6 @@ import 'package:conference_admin/features/imp-dates/domain/repositories/date_rep
 import 'package:conference_admin/features/imp-dates/domain/usecases/get_dates_uc.dart';
 import 'package:conference_admin/features/imp-dates/domain/usecases/update_dates_uc.dart';
 import 'package:conference_admin/features/imp-dates/presentation/bloc/imp_dates_bloc.dart';
-import 'package:conference_admin/features/pages/data/repositories/page_repo_impl.dart';
-import 'package:conference_admin/features/pages/domain/repositories/page_repo.dart';
-import 'package:conference_admin/features/pages/domain/usecases/get_page_uc.dart';
-import 'package:conference_admin/features/pages/domain/usecases/update_page_uc.dart';
-import 'package:conference_admin/features/pages/presentation/bloc/pages_bloc.dart';
 import 'package:get_it/get_it.dart';
 import 'package:conference_admin/features/login/data/repositories/login_repo_impl.dart';
 import 'package:conference_admin/features/login/domain/repositories/login_repo.dart';
@@ -76,7 +71,6 @@ Future<void> initializeDependencies() async {
   sl.registerSingleton<UsersRepo>(UsersRepoImpl(sl()));
   sl.registerSingleton<ScheduleRepo>(ScheduleRepoImpl());
   sl.registerSingleton<ArticleRepository>(ArticleRepoImpl(sl()));
-  sl.registerSingleton<PageRepo>(PageRepoImpl());
 
   //! Usecases
   //? About UseCases
@@ -86,9 +80,7 @@ Future<void> initializeDependencies() async {
   sl.registerSingleton<DeleteAboutPageUsecase>(DeleteAboutPageUsecase(sl()));
   sl.registerSingleton<CreateAboutPageUsecase>(CreateAboutPageUsecase(sl()));
 
-  //? Pages
-  sl.registerSingleton<GetPageUseCase>(GetPageUseCase(sl()));
-  sl.registerSingleton<UpdatePageUseCase>(UpdatePageUseCase(sl()));
+
 
   //? Home UseCases
   sl.registerSingleton<UpdateComponentUsecase>(UpdateComponentUsecase(sl()));
@@ -126,7 +118,6 @@ Future<void> initializeDependencies() async {
   //! Blocs
   //! About Blocs
   sl.registerFactory<AboutBloc>(() => AboutBloc(sl(), sl(), sl(), sl(), sl()));
-  sl.registerFactory<PagesBloc>(() => PagesBloc(sl(), sl()));
   sl.registerFactory<HomeBloc>(() => HomeBloc(sl(), sl(), sl(), sl(), sl()));
   sl.registerFactory<ImpDatesBloc>(() => ImpDatesBloc(sl(), sl()));
   sl.registerFactory<LoginBloc>(() => LoginBloc(sl(), sl(), sl()));