+++ admin-panel/lib/features/navbar/functions.dart
@@ -1,6 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:conference_admin/core/models/nav_bar_model.dart';
-import 'package:conference_admin/features/about/presentation/bloc/about_bloc.dart';
+import 'package:conference_admin/features/pages/presentation/bloc/about_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';