+++ admin-panel/lib/features/pages/domain/repositories/page_repo.dart
@@ -1,7 +1,0 @@
-import 'package:conference_admin/core/datastate/data_state.dart';
-import 'package:conference_admin/features/pages/data/models/pages_model.dart';
-
-abstract class PageRepo {
-  Future<DataState<PagesModel>> getPage(String id);
-  Future<DataState<void>> updatePage(PagesModel page);
-}