+++ admin-panel/lib/features/pages/domain/usecases/get_pages_uc.dart
@@ -1,8 +1,8 @@
 import 'package:conference_admin/core/const/no_params.dart';
 import 'package:conference_admin/core/datastate/data_state.dart';
 import 'package:conference_admin/core/usecase/usecase.dart';
-import 'package:conference_admin/features/about/domain/repositories/about_repo.dart';
-import 'package:conference_admin/features/pages/data/models/pages_model.dart';
+import 'package:conference_admin/features/pages/domain/repositories/about_repo.dart';
+import 'package:conference_admin/features/pages/data/models/page_model.dart';
 
 class GetAboutPagesUsecase extends UseCase<DataState<List<PagesModel>>, NoParams> {
   final AboutRepo aboutRepo;