+++ admin-panel/lib/features/pages/domain/usecases/update_component_uc.dart
@@ -1,7 +1,7 @@
 import 'package:conference_admin/core/datastate/data_state.dart';
 import 'package:conference_admin/core/usecase/usecase.dart';
-import 'package:conference_admin/features/about/domain/repositories/about_repo.dart';
-import 'package:conference_admin/features/pages/data/models/pages_model.dart';
+import 'package:conference_admin/features/pages/domain/repositories/about_repo.dart';
+import 'package:conference_admin/features/pages/data/models/page_model.dart';
 
 class UpdateAboutPageUsecase extends UseCase<DataState<void>, PagesModel> {
   final AboutRepo aboutRepo;